# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "youcare documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "enngsanae@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "organisator",
                                            "benevole",
                                            "admin"
                                        ],
                                        "example": "organisator"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    }
                }
            }
        },
        "/api/event": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all events",
                "operationId": "c41ead8dce94da3a165556310df2c2a6",
                "responses": {
                    "200": {
                        "description": "List of all events"
                    }
                }
            }
        },
        "/api/creatEvent": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "operationId": "b0cf4debacbdaae925bf4992b8ef9cc2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "date",
                                    "location",
                                    "type",
                                    "competences"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Event title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Event description"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Event location"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Event type"
                                    },
                                    "competences": {
                                        "type": "string",
                                        "example": "Event competences"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "401": {
                "description": "Unauthorized"
            },
            "422": {
                "description": "Validation error"
            }
        }
    }
}